more changes to snapshot script
Still need to get signing working.

Question: do we want to zip up all of the darcs data?

+++ scripts/snapshot.sh
@@ -4,14 +4,17 @@
 repo=/home/eric/haskell/kontrakcja
 filename=kontrakcja-snapshot-`date -u "+%Y%m%d%k%M%S%z"`
 cd /tmp
+echo "Copying repo"
 cp -r $repo $filename
 rm -rf $filename/_local
+echo "Zipping repo"
 tar zcf $filename.tar.gz $filename
 #sign with trustweaver
-twcert=/
-twcertpwd=abc
-twurl=http://
-curl -X POST -k --silent --show-error \
+twcert=/home/eric/haskell/kontrakcja/certs/credentials.pem
+twcertpwd=jhdaEo5LLejh
+twurl=https://tseiod.trustweaver.com/ts/svs.asmx
+echo "Signing with trustweaver"
+curl -X POST -k --verbose --show-error \
     --cert $twcert:$twcertpwd --cacert $twcert \
     --data-binary $filename.tar.gz \
     -H "Content-Type: text/xml; charset=UTF-8" \
@@ -19,9 +22,19 @@ curl -X POST -k --silent --show-error \
     -H "SOAPAction: http://www.trustweaver.com/tsswitch#Sign" \
     $twurl
 #push to amazon
-
+echo "Pushing to amazon"
+s3cmd --acl-private put $filename.tar.gz s3://skrivapa-snapshots
 #check hash from amazon
-#clean up
-rm -r $filename
-# don't delete file unless you can prove it's in amazon server
-#rm $filename.tar.gz
+echo "Checking amazon md5 sum"
+md5amazon=`s3cmd info s3://skrivapa-snapshots/$filename.tar.gz|grep MD5|awk "{print $3}"`
+echo $md5amazon
+md5local=`md5sum $filename.tar.gz | awk 'BEGIN { FS = " +" } ; { print $1 }'`
+if [$md5amazon==$md5local]
+then
+    echo "MD5 sum matches!"
+    #clean up
+    rm -r $filename
+    rm $filename.tar.gz
+    exit 0
+fi
+echo "Something went wrong. Try again."
